feat: Use spoke pool client lookback in dataworker
